PRJ: rename module
